Server/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
google-generativeai
dateparser
langchain
langchain-google-genai
boto3